requirements.txt
requests>=2.27.1
pymongo>=4.0.1
python-dateutil>=2.8.2